Pull out validation schema setup from form
